only expose necessary APIs in shm_alloc.h
Required-githooks: true
Skipped-githooks: codespell
